Simplify adif special fields
Make things that we want to parse into native types as property methods
instead of trying to calculate EVERYTHING we want and store them as
attrs.

Most cases you'll want to dig through AdifRecord.fields to access stuff,
but for example if you want a `date` to compare to another `date`, then
you can use the propery.